requirements.txt
Django==1.8.3
django-cors-headers==1.1.0
djangorestframework==3.1.3
djangorestframework-gis==0.9.1
jsonfield==1.0.3
-e git+https://github.com/inmagik/django-multi-gtfs
wsgiref==0.1.2
